pyproject.toml
[project]
name = "GENF"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "sibr-module[all]>=0.2.11",
    "streamlit>=1.51.0",
    "supabase>=2.24.0",
    "telethon>=1.41.2",
]
